+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 v 0.3.9
+- added channel confirable support IRC chat rooms
+- added IRC chat support
+- added fallback Podcast Index Item ID look up
 - added ability to create spit from scratch through split gui
 - Changed tip modal from lg to sm
 v 0.3.8